Added line space in ComputeGradientFeatures.py

+++ histomicstk/features/ComputeGradientFeatures.py
@@ -9,60 +9,45 @@ def ComputeGradientFeatures(im_label, im_intensity,
                             num_hist_bins=10, rprops=None):
     """
     Calculates gradient features from an intensity image.
-
     Parameters
     ----------
     im_label : array_like
         A labeled mask image wherein intensity of a pixel is the ID of the
         object it belongs to. Non-zero values are considered to be foreground
         objects.
-
     im_intensity : array_like
         Intensity image
-
     num_hist_bins: int, optional
         Number of bins used to computed the gradient histogram of an object.
         Histogram is used to energy and entropy features. Default is 10.
-
     rprops : output of skimage.measure.regionprops, optional
         rprops = skimage.measure.regionprops( im_label ). If rprops is not
         passed then it will be computed inside which will increase the
         computation time.
-
     Returns
     -------
     fdata: pandas.DataFrame
         A pandas dataframe containing the gradient features listed below for
         each object/label.
-
     Notes
     -----
     List of gradient features computed by this function:
-
     Gradient.Mag.Mean : float
         Mean of gradient data.
-
     Gradient.Mag.Std : float
         Standard deviation of gradient data.
-
     Gradient.Mag.Skewness : float
         Skewness of gradient data. Value is 0 when all values are equal.
-
     Gradient.Mag.Kurtosis : float
         Kurtosis of gradient data. Value is -3 when all values are equal.
-
     Gradient.Mag.HistEnergy : float
         Energy of the gradient magnitude histogram of object pixels
-
     Gradient.Mag.HistEnergy : float
         Entropy of the gradient magnitude histogram of object pixels.
-
     Gradient.Canny.Sum : float
         Sum of canny filtered gradient data.
-
     Gradient.Canny.Mean : float
         Mean of canny filtered gradient data.
-
     References
     ----------
     .. [1] Daniel Zwillinger and Stephen Kokoska. "CRC standard probability
@@ -126,7 +111,8 @@ def ComputeGradientFeatures(im_label, im_intensity,
         fdata.at[i, 'Gradient.Mag.HistEnergy'] = np.sum(prob**2)
 
         bw_canny = cannyG[rprops[i].coords[:, 0], rprops[i].coords[:, 1]]
-        canny_sum = np.sum(bw_canny)
+        canny_sum = np.sum(bw_canny).astype('float')
+        
         fdata.at[i, 'Gradient.Canny.Sum'] = canny_sum
 
         fdata.at[i, 'Gradient.Canny.Mean'] = canny_sum / len(pixelGradients)
